requirements.txt
pytesseract
opencv
pillow
